requirements.txt
async-lru==1.0.2
flake8==3.8.4
mccabe==0.6.1
pyaes==1.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
Pyrogram==1.1.13
PySocks==1.7.1
python-dotenv==0.15.0
